+++ vite.config.ts
@@ -10,7 +10,7 @@ export default defineConfig({
     tailwindcss(),
     VitePWA({
       injectRegister: "auto",
-      registerType: "autoUpdate",
+      registerType: "prompt",
       manifest: {
         name: "Emailable",
         short_name: "Emailable",